requirements.txt
delta-sharing==1.4.1
pandas==3.0.1
python-dotenv==1.2.2
requests==2.32.5
pandasql==0.7.3
anthropic==0.84.0
beautifulsoup4==4.14.3
lxml==6.0.2
python-docx==1.2.0
python-pptx==1.0.2
